Fix incorrect bit shifts and incorrect ChannelGetLevelLeft/ChannelGetLevelRight returns (#40)
* Fix HiWord / HiDword performing arithmetic right-shifts instead of logical right-shifts
* Make ChannelGetLevelLeft/ChannelGetLevelRight return -1 if an error has occurred

BASS explicitly returns -1 if an error has occurred (http://www.bass.radio42.com/help/html/5f031dfc-effe-35a9-c60d-21856e15a59e.htm). We can just pass this value back to the user to signal that they should check Bass.LastError.

+++ src/Bass/Shared/BitHelper.cs
@@ -10,7 +10,7 @@ namespace ManagedBass
         /// <summary>
         /// The return value is the high-order double word of the specified value.
         /// </summary>
-        public static long HiDword(this long DWord) => DWord >> 32;
+        public static long HiDword(this long DWord) => (long)((ulong)DWord >> 32);
 
         /// <summary>
         /// The return value is the low-order word of the specified value.
@@ -20,7 +20,7 @@ namespace ManagedBass
         /// <summary>
         /// The return value is the high-order word of the specified value.
         /// </summary>
-        public static int HiWord(this int DWord) => DWord >> 16;
+        public static int HiWord(this int DWord) => (int)((uint)DWord >> 16);
 
         /// <summary>
         /// The return value is the low-order word of the specified value.
